docker_database_interface/requirements.txt
psycopg2-binary
uvicorn
fastapi
requests
